// frontend/src/api/generated/models/codeLinkRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { CodeLinkKind } from './codeLinkKind';
import type { PullRequestState } from './pullRequestState';
import type { RepositoryLinkedItem } from './repositoryLinkedItem';

export interface CodeLinkRead {
  id: number;
  kind: CodeLinkKind;
  external_id: string;
  title: string | null;
  url: string;
  state: PullRequestState | null;
  author_name: string | null;
  repository: RepositoryLinkedItem;
  created_at: string;
  updated_at: string;
}